+++ parley_engine/src/lib.rs
@@ -107,7 +107,6 @@ extern crate alloc;
 mod analysis;
 mod analyzer;
 pub mod bidi;
-pub mod break_overrides;
 mod glyph;
 pub mod itemize;
 mod lru_cache;